Tests/LibWeb/Text/expected/canvas/pattern-from-video.txt

null
PASS (didn't throw!)
